Move settings-parsing helpers from state.py into settings.py
_camel_to_snake, _normalise_formatter_settings, _KNOWN_TCL_LSP_SECTIONS,
_KNOWN_TCL_LSP_TOPLEVEL, and _extract_tcl_lsp_settings belong with the
settings parsing logic, not with singleton state. Remove import re from
state.py (now unused); update workspace_init.py and test_server_config.py
to import the helpers from lsp.settings.
