+++ tests/test_server_config.py
@@ -10,7 +10,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
 import lsp.settings as _lsp_settings
 import lsp.state as _lsp_state
 from lsp.feature_config import FeatureConfig
-from lsp.state import _extract_tcl_lsp_settings
+from lsp.settings import _extract_tcl_lsp_settings
 
 
 def test_optimiser_filters_include_o103_o104():
